+++ vitest.config.ts
@@ -1,7 +1,13 @@
 // vitest.config.ts
 import { defineConfig } from 'vitest/config'
+import path from 'node:path'
 
 export default defineConfig({
+  resolve: {
+    alias: {
+      '@': path.resolve(__dirname, './src')
+    }
+  },
   test: {
     environment: 'happy-dom',
     coverage: {